requirements.txt
astroid==2.11.2
black==22.3.0
click==8.0.3
dill==0.3.4
Flask==2.0.2
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
lazy-object-proxy==1.7.1
MarkupSafe==2.0.1
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pylint==2.13.4
python-dotenv==0.19.2
tomli==2.0.1
typing-extensions==4.1.1
Werkzeug==2.0.2
wrapt==1.14.0
